Bugs solved

+++ bursts_plot.py
@@ -156,7 +156,7 @@ def plot(DS, subplot_spec, fig, extent=None, ncols=1, nrows=1, t_scrunch=1., f_s
     extent[0] = - width / 2.
     extent[1] = width / 2.
     components_ms = components / t_scrunch * res_t
-    components = np.int(components / t_scrunch)
+    components = (components / t_scrunch).astype(int)
     if width: 
       smooth_DS = smooth_DS[:, int(prof.argmax() - np.ceil(width / 2. / res_t)) : int(prof.argmax() + np.ceil(width / 2. / res_t))]
       components_ms -= peak_ms
